+++ packages/core/src/types/curi.ts
@@ -1,9 +1,15 @@
-import { History, HickoryLocation, Action } from "@hickory/root";
+import {
+  History,
+  LocationDetails,
+  HickoryLocation,
+  PartialLocation,
+  Action
+} from "@hickory/root";
 import { PathFunctionOptions } from "path-to-regexp";
 
 import { Addon, Addons } from "./addon";
 import { RouteDescriptor } from "./route";
-import { Response } from "./response";
+import { Response, Params } from "./response";
 
 export interface Navigation {
   action: Action;
@@ -47,6 +53,11 @@ export interface CurrentResponse {
   navigation: Navigation;
 }
 
+export interface LocationProps extends LocationDetails {
+  name: string;
+  params?: Params;
+}
+
 export interface CuriRouter {
   refresh: (routeArray: Array<RouteDescriptor>) => void;
   respond: (
@@ -56,4 +67,5 @@ export interface CuriRouter {
   addons: Addons;
   history: History;
   current(): CurrentResponse;
+  createLocation(options: LocationProps): PartialLocation;
 }